omxh264enc: Localize ref-frames property
Specification of nRefFrames of R-Car is different from
Khronos specification, and 0 can be specified for R-Car.

Actual limit for nRefFrames for R-Car is from 0 to 2,
but set G_MAXUINT for upper limit for backward compatibility.
If specify larger than 2, component automatically modify
according to other parameters.

+++ omx/gstomxh264enc.h
@@ -55,7 +55,7 @@ struct _GstOMXH264Enc
   guint32 entropy_mode;
   gboolean constrained_intra_prediction;
   guint32 loop_filter_mode;
-  guint8 ref_frames;
+  guint32 ref_frames;
 
   GList *headers;
 };
